requirements.txt
Django>=4.1.5
psycopg>=2.9.5
psycopg[binary]
pillow>=9.4.0
djangorestframework>=3.14.0
drf-nested-routers>=0.93.4
djoser>=2.1.0
django-phonenumber-field[phonenumberslite]>=7.0.2
drf-yasg>=1.21.4
gunicorn>=20.1.0
dj-database-url>=1.2.0
django-environ>=0.9.0
whitenoise>=6.3.0
django-debug-toolbar>=3.8.1
django-filter>=22.1